3-step_model_final/stage3_final/stage_3_cnn/stage3-cnn-explainer/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
tensorflow-cpu==2.16.1
opencv-python-headless==4.10.0.84
